Add support for "ALTER TABLE t COLUMN c INT FIRST", and fix bug when primary key columns move.
Id columns always move when adding a first column, but they could also move when adding a column using AFTER, or when dropping a column if the primary key isn't the first column.

Probably a rare case, but the result if it happens would be that the column order in the internal state doesn't match the database anymore, which is bad.